Rename pubProvidedSystem.js to pubProvidedIdSystem.js (#5861)
* Rename pubProvidedSystem.js to pubProvidedIdSystem.js

* Update userId_spec.js
